Nest analytics into AppHead component
