+++ components/AppHead/index.js
@@ -1,0 +1,3 @@
+import AppHead from './AppHead';
+
+export default AppHead;
